refactoring: ckbtc receive modal to prepare receive on accounts
